+++ bomeba0/.gitignore
@@ -1,0 +1,2 @@
+.ipynb_checkpoints
+__pycache__/
