+++ bomeba0/__init__.py
@@ -1,4 +1,5 @@
 from .molecules import *
 from .visualization.plots import *
 from .superposition import *
+from .utils import *
 from .external import *
