+++ bomeba0/energy/__init__.py
@@ -1,0 +1,1 @@
+from .ff import *
