+++ bomeba0/pdbs/1d3z0.pdb
@@ -6,9 +6,7 @@ ATOM      5  CB  MET A   1      51.976 -87.958   9.485  1.00 13.77           C
 ATOM      6  CG  MET A   1      52.864 -87.131   8.557  1.00 16.29           C  
 ATOM      7  SD  MET A   1      53.355 -85.606   9.403  1.00 17.17           S  
 ATOM      8  CE  MET A   1      53.359 -84.524   7.954  1.00 16.11           C  
-ATOM      9  H1  MET A   1      52.721 -91.019   8.328  1.00  0.00           H  
-ATOM     10  H2  MET A   1      53.581 -89.933   9.311  1.00  0.00           H  
-ATOM     11  H3  MET A   1      53.352 -89.592   7.663  1.00  0.00           H  
+ATOM      9  H   MET A   1      52.721 -91.019   8.328  1.00  0.00           H  
 ATOM     12  HA  MET A   1      51.073 -89.904   9.519  1.00  0.00           H  
 ATOM     13  HB2 MET A   1      51.058 -87.421   9.674  1.00  0.00           H  
 ATOM     14  HB3 MET A   1      52.492 -88.123  10.416  1.00  0.00           H  
@@ -1225,7 +1223,6 @@ ATOM   1224  N   GLY A  76      46.647 -71.781 -23.672  0.25 36.05           N
 ATOM   1225  CA  GLY A  76      45.367 -71.266 -23.201  0.25 36.19           C  
 ATOM   1226  C   GLY A  76      44.610 -70.574 -24.328  0.25 36.20           C  
 ATOM   1227  O   GLY A  76      44.420 -69.372 -24.237  1.00  0.00           O  
-ATOM   1228  OXT GLY A  76      44.230 -71.255 -25.266  1.00  0.00           O  
 ATOM   1229  H   GLY A  76      46.825 -71.824 -24.635  1.00  0.00           H  
 ATOM   1230  HA2 GLY A  76      44.773 -72.086 -22.824  1.00  0.00           H  
 ATOM   1231  HA3 GLY A  76      45.541 -70.558 -22.406  1.00  0.00           H
