+++ bomeba0/superposition/__init__.py
@@ -1,0 +1,2 @@
+from .super import rmsd_fit
+
